Stop drift detecting cluster classic property
https://github.com/aws-cloudformation/cloudformation-coverage-roadmap/issues/1204

Classic property indicates whether the resize operation is using the classic resize.
It's added to ease cloudformation template to indicate classic resize.
describe-clusters used in drift will not indicate this property and hence this
property once used in updating should be removed.

This commit adds this classic to "writeOnlyProperties". CFN drift
detection doesn't check on writeOnlyProperties.