Prevent cluster drift detection on SnapshotIdentifier
Right now if users provide SnapshotIdentifier on the Cluster in CFN
template, after its creation, drift detection will display the cluster's
status as drifted. In this case, we restore the cluster from the
snapshot provided, and the snapshot identifier should not be considered
as part of the drift detection.

Adding it to writeOnly properties will prevent this from happening