Wait 10 seconds after modify-cluster call
After the modify-cluster call, it takes 1 or more seconds for the ACTIVE cluster to be changed to MODIFYING,
we used to call stabilizer right after the modify-cluster call, and stabilizer immediately returns that the cluster
is ACTIVE, because the cluster didn't get a chance to be changed to be MODIFYING, which incorrectly tells
CFN that the modify-cluster is done.

We wait for 10 seconds just to be on the safe side, it's longer than we need most of the time, but 10 seconds
is acceptable because the modify-cluster usually takes minutes for the cluster to be ACTIVE again.