CFN support for MultiAZ (#150)
* CFN support for MultiAZ

The changes provide CFN support for MultiAZ redshift cluster. It allows us to create/modify MultiAZ redshift cluster using the cloudformation template. The changes in this request allow us to use create-cluster, modify-cluster, restore-from-cluster-snapshot and failover-primary-compute APIs for MultiAZ redshift cluster.

We are adding a new boolean parameter "multiAZ" to differentiate between a Single-AZ and MultiAZ cluster and the parameter is used while setting up the CreateClusterRequest, ModifyClusterRequest and RestoreFromClusterSnapshotRequest.

FailoverPrimaryCompute is a new API which will be supported for MultiAZ redshift cluster and can be used with the help of resource-action attribute which also supports pause-cluster and resume-cluster. The cluster has to be a MultiAZ redshift cluster to trigger FailoverPrimaryCompute API. If we try to trigger failover-primary-compute on a single-az cluster, redshift returns the call with an InvalidClusterState exception and a prompt message.

Manual testing has been done on Cloudformation using Cloudformation template for all the changes.

---------

Co-authored-by: Akarsh Dang <dangakar@amazon.com>