System level tags are not supported in Redshift
We dont support adding system tags to of the form aws:<composite-service>:<key> = <value>. Tagris recommends you allow them to do that but it is up to the service. The composite service does not have a way to add a system tag on a resource from Redshift.

Another options would be whitelisted by tAgris to allow the system tags. For now, we are removing the support for the system tags