System level tags are not supported in Redshift  (#155)
* Add Stack tags support to create-cluster call

The create-cluster can be tagged with only resource-level tags, stack level tags are not added as the part of this process to the resource.
Stack Level Tags: are tags specified in the AWS console, they should be shared by all resources created inside that stack.
Resource Level Tags: are tags specified in CloudFormation templates.
CloudFormation System Tags: In addition to any tags you define, CloudFormation automatically creates the following system tags with the prefix aws:
* aws:cloudformation:logical-id
* aws:cloudformation:stack-id
* aws:cloudformation:stack-name

* System level tags are not supported in Redshift

We dont support adding system tags to of the form aws:<composite-service>:<key> = <value>. Tagris recommends you allow them to do that but it is up to the service. The composite service does not have a way to add a system tag on a resource from Redshift.

Another options would be whitelisted by tAgris to allow the system tags. For now, we are removing the support for the system tags