Cloudformation support for Secrets Manager Integration (#166)
The changes provide Cloudformation support for Secrets Manager integration with Redshift clusters. With this new feature, customers can opt in to store their cluster's admin credentials in a service linked secret in Secrets Manager. It allows us to create/modify/restore Redshift clusters with Secrets Manager support using the Cloudformation template. The changes in this request allow us to use create-cluster, modify-cluster, restore-from-cluster-snapshot APIs for Redshift clusters when opting in to this feature.

We are adding a new boolean parameter "ManageMasterPassword" to allow customers to opt in to this feature and another parameter "MasterPasswordSecretKmsKeyId" allows customers to specify the key ID of the KMS key in the customer account which will be used to encrypt the cluster secret. These parameters can be used while setting CreateClusterRequest, ModifyClusterRequest and RestoreFromClusterSnapshotRequest. The response of these requests will return the "MasterPasswordSecretArn" when the cluster is opted in to this feature.