Refactored cluster parameter group name default value logic when group name is not provided
This CR refactors the logic related to updating the default group name.
We access the input model object directly instead of through callback
functions. The CR will fix failures in cluster paramater group creation
when group name is not provided.

Tested locally when group name is provided and group name is not provided
Testing commands:

        1. sample json
        {
        "desiredResourceState": {
        "Description": "Redshift ClusterParameterGroup Contract Test V2",
        "ParameterGroupName": "vramanet-cpg",
        "ParameterGroupFamily": "redshift-1.0",
        "Parameters": [
        {
        "ParameterName": "require_ssl",
        "ParameterValue": "true"
        }
        ]
        },
        "previousResourceState": {},
        "logicalResourceIdentifier": "vramanet-cpg"
        }
        2. sam local start-lambda
        3. cfn invoke resource CREATE create-clusterparametergroup.json