Check track name changes before modify-cluster
Before this fix, if cx doesn't provide a track name in the CFN template,
we'll call the create-cluster w/o the track name, and create-cluster by
default creats a cluster w/ "current" track. When cx does add "current"
track to the template, our code thinks it needs an update and calls
modify-cluster w/ "current" track, and modify-cluster throws exception
that the cluster is already on "current" track.

We should have handled this and not call modify-cluster when track name
changes from null to "current".

This commit checks the previousModel's track setup and calls
modify-cluster accordingly